+++ package.json
@@ -9,13 +9,12 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@prisma/client": "^5.15.0",
+    "@prisma/client": "^5.15.1",
     "axios": "^1.7.2",
     "crypto-js": "^4.2.0",
     "mysql": "^2.18.1",
     "next": "14.2.4",
     "next-auth": "^4.24.7",
-    "prisma": "^5.15.0",
     "react": "^18",
     "react-dom": "^18",
     "sass": "^1.77.6"
@@ -28,6 +27,7 @@
     "@types/react-dom": "^18",
     "eslint": "^8",
     "eslint-config-next": "14.2.4",
+    "prisma": "^5.15.1",
     "typescript": "^5.4.5"
   },
   "type": "module"
