+++ tsconfig.json
@@ -17,8 +17,11 @@
         "name": "next"
       }
     ],
+    "baseUrl": "./",
     "paths": {
-      "@/*": ["./*"]
+      "@/*": ["./*"],
+      "@/app/*": ["app/*"],
+      "@/components/*": ["app/components/*"]
     }
   },
   "include": [
